second assignment

+++ src/main/java/project/spring/model/business/service/AssignmentService.java
@@ -8,17 +8,21 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import project.spring.model.business.model.Assignment;
+import project.spring.model.business.model.Lab;
 import project.spring.model.dal.dto.AssignmentDto;
 import project.spring.model.dal.repository.IAssignmentRepository;
+import project.spring.model.dal.repository.ILabRepository;
 
 @Service
 public class AssignmentService implements IAssignmentService {
 
 	private final IAssignmentRepository assignmentRepository;
+	private final ILabRepository laboratoryRepository;
 
 	@Autowired
-	public AssignmentService(IAssignmentRepository assignmentRepository) {
+	public AssignmentService(IAssignmentRepository assignmentRepository, ILabRepository laboratoryRepository) {
 		this.assignmentRepository = assignmentRepository;
+		this.laboratoryRepository = laboratoryRepository;
 	}
 
 	@Override
@@ -33,31 +37,61 @@ public class AssignmentService implements IAssignmentService {
 	@Override
 	public AssignmentDto getAssignmentById(Long id) {
 		Assignment assignment = assignmentRepository.findOne(id);
-		return mapDto(assignment);
+		if (assignment != null) {
+			return mapDto(assignment);
+		} else {
+			return null;
+		}
 	}
 
 	@Override
-	public void saveAssignment(AssignmentDto assignment) {
+	public boolean saveAssignment(AssignmentDto assignment, Long laboratoryId) {
 		Assignment assignmentToSave = this.map(assignment);
-		assignmentRepository.save(assignmentToSave);
+		Lab lab = laboratoryRepository.findOne(laboratoryId);
+		if (lab != null) {
+			assignmentToSave.setLab(lab);
+			assignmentRepository.save(assignmentToSave);
+			return true;
+		}
+		return false;
 
 	}
 
 	@Override
-	public void updateAssignment(Long id, AssignmentDto assignment) {
+	public boolean updateAssignment(Long id, AssignmentDto assignment) {
 		Assignment assignmentToUpdate = assignmentRepository.findOne(id);
 		if (assignmentToUpdate != null) {
 			assignmentToUpdate = map(assignment);
 			assignmentToUpdate.setAssignmentId(id);
 			assignmentRepository.save(assignmentToUpdate);
+			return true;
 		}
+		return false;
 
 	}
 
 	@Override
-	public void deleteAssignment(Long id) {
-		assignmentRepository.delete(id);
+	public boolean deleteAssignment(Long id) {
+		Assignment assignment = assignmentRepository.findOne(id);
+		if (assignment != null) {
+			assignmentRepository.delete(id);
+			return true;
+		}
+		return false;
 
+	}
+
+	@Override
+	public List<AssignmentDto> getAllAssignmentsForLab(Long laboratoryId) {
+		List<AssignmentDto> assignmentsDto = new ArrayList<AssignmentDto>();
+		List<Assignment> assignments = new ArrayList<Assignment>();
+		Lab lab = laboratoryRepository.findOne(laboratoryId);
+		if (lab != null) {
+			assignmentRepository.findByLabLaboratoryId(laboratoryId).forEach(assignments::add);
+			assignmentsDto = assignments.stream().map(s -> this.mapDto(s)).collect(Collectors.toList());
+			return assignmentsDto;
+		}
+		return null;
 	}
 
 	public AssignmentDto mapDto(Assignment assignment) {
