+++ src/main/java/project/spring/model/business/service/TeacherService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import project.spring.model.business.model.Teacher;
 import project.spring.model.dal.dto.TeacherDto;
 import project.spring.model.dal.repository.ITeacherRepository;
+import project.spring.utils.PasswordEncryptor;
 
 @Service
 public class TeacherService implements ITeacherService {
@@ -33,40 +34,70 @@ public class TeacherService implements ITeacherService {
 	@Override
 	public TeacherDto getTeacherById(Long id) {
 		Teacher teacher = teacherRepository.findOne(id);
-		return mapDto(teacher);
-	}
-
-	@Override
-	public void saveTeacher(TeacherDto teacher) {
-		Teacher teacherToSave = this.map(teacher);
-		teacherRepository.save(teacherToSave);
-
-	}
-
-	@Override
-	public void updateTeacher(Long id, TeacherDto teacher) {
-		Teacher teacherToUpdate = teacherRepository.findOne(id);
-		if (teacherToUpdate != null) {
-			teacherToUpdate = map(teacher);
-			teacherToUpdate.setTeacherId(id);
-			teacherRepository.save(teacherToUpdate);
+		if (teacher != null) {
+			return mapDto(teacher);
+		} else {
+			return null;
 		}
 
 	}
 
 	@Override
-	public void deleteTeacher(Long id) {
-		teacherRepository.delete(id);
+	public boolean saveTeacher(TeacherDto teacher) {
+		Teacher teacherToSave = this.map(teacher);
+		if (teacherRepository.findByUsername(teacherToSave.getUsername()) != null) {
+			return false;
+		} else {
+			teacherToSave.setPassword(PasswordEncryptor.setPasswordEncrypt(teacherToSave.getPassword()));
+			teacherRepository.save(teacherToSave);
+			return true;
+		}
+	}
+
+	@Override
+	public boolean updateTeacher(Long id, TeacherDto teacher) {
+		Teacher teacherToUpdate = teacherRepository.findOne(id);
+		if (teacherToUpdate != null) {
+			teacherToUpdate = map(teacher);
+			teacherToUpdate.setTeacherId(id);
+			teacherToUpdate.setPassword(PasswordEncryptor.setPasswordEncrypt(teacherToUpdate.getPassword()));
+			teacherRepository.save(teacherToUpdate);
+			return true;
+		}
+		return false;
 
 	}
 
+	@Override
+	public boolean deleteTeacher(Long id) {
+		Teacher teacherToDelete = teacherRepository.findOne(id);
+		if (teacherToDelete != null) {
+			teacherRepository.delete(id);
+			return true;
+		} else {
+			return false;
+		}
+
+	}
+
+	@Override
+	public TeacherDto login(String username, String password) {
+		Teacher teacher = teacherRepository.findByUsername(username);
+		if (teacher != null) {
+			if (teacher.getPassword().equals(PasswordEncryptor.setPasswordEncrypt(password))) {
+				return mapDto(teacher);
+			}
+		}
+		return null;
+	}
+
 	public TeacherDto mapDto(Teacher teacher) {
-		TeacherDto teacherDto = new TeacherDto(teacher.getEmail(), teacher.getPassword());
+		TeacherDto teacherDto = new TeacherDto(teacher.getUsername(), teacher.getPassword());
 		return teacherDto;
 	}
 
 	public Teacher map(TeacherDto teacher) {
-		Teacher tech = new Teacher(teacher.getEmail(), teacher.getPassword());
+		Teacher tech = new Teacher(teacher.getUsername(), teacher.getPassword());
 		return tech;
 	}
